Rename ListBufferByteTag
