+++ storage/ListBuffer.h
@@ -6,7 +6,7 @@
 #include "ListByteBuffer.h"
 #include "ListElementViewBuffer.h"
 
-#include "ListBufferByteTag.h"
+#include "ListBufferTypeTag.h"
 
 #include "ListElementView.h"
 #include "ListView.h"
